Проведен рефакторинг страницы ServicesView
