--- src/components/Dialogs/MakeService.vue
@@ -1,12 +1,0 @@
-<template>
-    <div>
-        
-    </div>
-</template>
-<script setup>
-
-</script>
-
-<style scoped>
-    
-</style>
