--- src/components/Dialogs/MoreService.vue
@@ -1,14 +1,0 @@
-<template>
-    <div>
-
-    </div>
-</template>
-<script setup>
-import Accordion from 'primevue/accordion';
-import AccordionPanel from 'primevue/accordionpanel';
-import AccordionHeader from 'primevue/accordionheader';
-import AccordionContent from 'primevue/accordioncontent';
-
-</script>
-
-<style scoped></style>
